LPS-196947 Inline PanelAppMyAccountPermissions methods into CompanyModelListener, change ServiceTracker to ServiceTrackerMap, so it becomes lazy, change PanelAppServiceTrackerCustomizer interface implementation to EagerServiceTrackerCustomizer to initialize the tracker when it is open
